test: extract shared records helpers
