+++ tests/tests_integration/test_api/test_data_modeling/test_records.py
@@ -1,15 +1,10 @@
 from __future__ import annotations
 
-import time
-import uuid
-from collections.abc import Callable, Iterator
-from dataclasses import dataclass
-from datetime import datetime, timedelta, timezone
+from collections.abc import Iterator
 
 import pytest
 
 from cognite.client import CogniteClient
-from cognite.client.data_classes import filters
 from cognite.client.data_classes.data_modeling import (
     Boolean,
     ContainerApply,
@@ -18,63 +13,20 @@ from cognite.client.data_classes.data_modeling import (
     Space,
     Text,
 )
-from cognite.client.data_classes.data_modeling.aggregates import Average, Count, MetricResult
-from cognite.client.data_classes.data_modeling.records import (
-    RecordContainerId,
-    RecordId,
-    RecordSource,
-    RecordSourceSelector,
-    RecordWrite,
-    TimeRange,
+from cognite.client.data_classes.data_modeling.records import RecordContainerId, RecordSourceSelector
+from cognite.client.data_classes.data_modeling.streams import Stream
+from tests.records_helpers import (
+    RecordBatch,
+    assert_eventually,
+    assert_filtered_records,
+    assert_record_aggregates,
+    assert_sync_records,
+    get_or_create_mutable_stream,
+    record_batch,
+    upsert_and_assert_record,
 )
-from cognite.client.data_classes.data_modeling.streams import (
-    Stream,
-    StreamTemplate,
-    StreamWrite,
-    StreamWriteSettings,
-)
-from cognite.client.exceptions import CogniteAPIError
-from cognite.client.utils._retry import Backoff
 
-# Deleted streams are soft deleted, and their external IDs stay reserved for a couple of weeks
-# before they can be reused. Combined with the low quota on active streams per project, these
-# tests use one fixed stream (created only if missing) rather than one per run.
-STREAM_EXTERNAL_ID = "sdk_test_mutable_stream"
 CONTAINER_EXTERNAL_ID = "PythonSdkIntegrationTestRecords"
-CONSISTENCY_TIMEOUT = 60.0
-
-
-def an_hour_ago() -> str:
-    """A lower bound for 'lastUpdatedTime'.
-
-    TimeRange forwards its bounds to the API untouched, and the API only accepts ISO-8601 there -
-    not the "1h-ago" shorthand that the rest of the SDK understands.
-    """
-    return (datetime.now(timezone.utc) - timedelta(hours=1)).strftime("%Y-%m-%dT%H:%M:%SZ")
-
-
-def assert_eventually(assertion: Callable[[], None], *, deadline: float | None = None) -> None:
-    """Poll only assertion failures, retaining the final failure as the timeout's cause."""
-    if deadline is None:
-        deadline = time.monotonic() + CONSISTENCY_TIMEOUT
-    wait = Backoff(max_wait=4, min_wait=0.25)
-    while True:
-        try:
-            assertion()
-            return
-        except AssertionError as error:
-            remaining = deadline - time.monotonic()
-            if remaining <= 0:
-                name = getattr(assertion, "__name__", type(assertion).__name__)
-                raise AssertionError(f"Timed out waiting for {name}: {error}") from error
-            time.sleep(min(next(wait), remaining))
-
-
-@dataclass
-class RecordBatch:
-    records: list[RecordWrite]
-    filter: filters.Equals
-    cursor: str
 
 
 @pytest.fixture(scope="session")
@@ -97,29 +49,7 @@ def record_container(cognite_client: CogniteClient, integration_test_space: Spac
 
 @pytest.fixture(scope="session")
 def mutable_stream(cognite_client: CogniteClient) -> Stream:
-    deadline = time.monotonic() + CONSISTENCY_TIMEOUT
-    stream = cognite_client.data_modeling.streams.retrieve(STREAM_EXTERNAL_ID)
-    if stream is None:
-        try:
-            stream = cognite_client.data_modeling.streams.create(
-                StreamWrite(
-                    external_id=STREAM_EXTERNAL_ID,
-                    settings=StreamWriteSettings(template=StreamTemplate(name="BasicLiveData")),
-                )
-            )
-        except CogniteAPIError as error:
-            if error.code != 409:
-                raise
-
-            def concurrently_created_stream_is_visible() -> None:
-                nonlocal stream
-                stream = cognite_client.data_modeling.streams.retrieve(STREAM_EXTERNAL_ID)
-                assert stream is not None, f"Conflicting stream {STREAM_EXTERNAL_ID!r} is not retrievable"
-
-            assert_eventually(concurrently_created_stream_is_visible, deadline=deadline)
-    assert stream is not None
-    assert stream.type == "Mutable", f"Shared stream {STREAM_EXTERNAL_ID!r} must be mutable"
-    return stream
+    return get_or_create_mutable_stream(cognite_client)
 
 
 @pytest.fixture(scope="session")
@@ -138,51 +68,8 @@ def ingested_records(
     mutable_stream: Stream,
     container_ref: RecordContainerId,
 ) -> Iterator[RecordBatch]:
-    tag = uuid.uuid4().hex
-    records = [
-        RecordWrite(
-            space=container_ref.space,
-            external_id=f"{tag}-{i}",
-            sources=[
-                RecordSource(
-                    source=container_ref,
-                    properties={"name": tag, "value": float(i), "processed": i % 2 == 0},
-                )
-            ],
-        )
-        for i in range(3)
-    ]
-    tagged = filters.Equals(property=[container_ref.space, container_ref.external_id, "name"], value=tag)
-    # Capture a stable starting position before writing. Reinitializing a relative window on
-    # each retry can move past our records when CI is slow. The UUID has never been used before.
-    cursor = next(
-        cognite_client.data_modeling.records.sync(
-            stream_id=mutable_stream.external_id,
-            initialize_cursor="1m-ago",
-            filter=tagged,
-        )
-    ).cursor
-    assert cursor is not None
-    batch = RecordBatch(records, tagged, cursor)
-
-    def all_records_are_queryable() -> None:
-        result = cognite_client.data_modeling.records.filter(
-            stream_id=mutable_stream.external_id,
-            last_updated_time=TimeRange(gt=an_hour_ago()),
-            filter=batch.filter,
-            limit=len(records) + 1,
-        )
-        assert {record.as_id() for record in result} == {record.as_id() for record in records}
-        assert len(result) == len(records)
-
-    try:
-        cognite_client.data_modeling.records.ingest(records, stream_id=mutable_stream.external_id)
-        assert_eventually(all_records_are_queryable)
+    with record_batch(cognite_client, mutable_stream, container_ref) as batch:
         yield batch
-    finally:
-        cognite_client.data_modeling.records.delete(
-            [record.as_id() for record in records], stream_id=mutable_stream.external_id
-        )
 
 
 class TestRecordsIntegration:
@@ -194,18 +81,7 @@ class TestRecordsIntegration:
         sources: list[RecordSourceSelector],
         ingested_records: RecordBatch,
     ) -> None:
-        def filtered_records_are_visible() -> None:
-            result = cognite_client.data_modeling.records.filter(
-                stream_id=mutable_stream.external_id,
-                last_updated_time=TimeRange(gt=an_hour_ago()),
-                sources=sources,
-                filter=ingested_records.filter,
-                limit=10,
-            )
-            assert len(result) == len(ingested_records.records)
-            assert {record.external_id for record in result} == {r.external_id for r in ingested_records.records}
-
-        assert_eventually(filtered_records_are_visible)
+        assert_filtered_records(cognite_client, mutable_stream, container_ref, sources, ingested_records)
 
     def test_aggregate_over_ingested_records(
         self,
@@ -214,21 +90,7 @@ class TestRecordsIntegration:
         container_ref: RecordContainerId,
         ingested_records: RecordBatch,
     ) -> None:
-        value = [container_ref.space, container_ref.external_id, "value"]
-
-        def aggregates_are_visible() -> None:
-            result = cognite_client.data_modeling.records.aggregate(
-                {"total": Count(), "avg_value": Average(value)},
-                stream_id=mutable_stream.external_id,
-                last_updated_time=TimeRange(gt=an_hour_ago()),
-                filter=ingested_records.filter,
-            )
-            total, avg_value = result["total"], result["avg_value"]
-            assert isinstance(total, MetricResult) and isinstance(avg_value, MetricResult)
-            assert total.value == len(ingested_records.records)
-            assert avg_value.value == pytest.approx(1.0)  # mean of 0.0, 1.0, 2.0
-
-        assert_eventually(aggregates_are_visible)
+        assert_record_aggregates(cognite_client, mutable_stream, container_ref, ingested_records)
 
     @pytest.mark.parametrize("chunk_size", [1000, 2])
     def test_sync_iterates_until_feed_exhausted(
@@ -239,29 +101,7 @@ class TestRecordsIntegration:
         ingested_records: RecordBatch,
         chunk_size: int,
     ) -> None:
-        """Accept arbitrary page boundaries and stop only when has_next is false."""
-        deadline = time.monotonic() + CONSISTENCY_TIMEOUT
-
-        def feed_yields_every_record() -> None:
-            seen: set[RecordId] = set()
-            exhausted = False
-            for page in cognite_client.data_modeling.records.sync(
-                stream_id=mutable_stream.external_id,
-                cursor=ingested_records.cursor,
-                sources=sources,
-                filter=ingested_records.filter,
-                chunk_size=chunk_size,
-            ):
-                assert time.monotonic() < deadline, "sync did not exhaust the feed before the deadline"
-                assert not exhausted, "sync yielded another page after has_next=False"
-                assert page.cursor is not None
-                assert len(page) <= chunk_size
-                seen.update(record.as_id() for record in page)
-                exhausted = page.has_next is False
-            assert exhausted, "sync ended without has_next=False"
-            assert seen == {record.as_id() for record in ingested_records.records}
-
-        assert_eventually(feed_yields_every_record, deadline=deadline)
+        assert_sync_records(cognite_client, mutable_stream, sources, ingested_records, chunk_size)
 
     def test_upsert_replaces_record(
         self,
@@ -271,33 +111,7 @@ class TestRecordsIntegration:
         sources: list[RecordSourceSelector],
         ingested_records: RecordBatch,
     ) -> None:
-        target = ingested_records.records[0]
-        properties = {**target.sources[0].properties, "value": 99.0}
-        cognite_client.data_modeling.records.upsert(
-            RecordWrite(
-                space=target.space,
-                external_id=target.external_id,
-                sources=[RecordSource(source=container_ref, properties=properties)],
-            ),
-            stream_id=mutable_stream.external_id,
-        )
-
-        def replacement_is_queryable() -> None:
-            result = cognite_client.data_modeling.records.filter(
-                stream_id=mutable_stream.external_id,
-                last_updated_time=TimeRange(gt=an_hour_ago()),
-                sources=sources,
-                filter=filters.And(
-                    ingested_records.filter,
-                    filters.Equals(property=[container_ref.space, container_ref.external_id, "value"], value=99.0),
-                ),
-                limit=10,
-            )
-            assert [record.as_id() for record in result] == [target.as_id()]
-            assert result[0].properties is not None
-            assert result[0].properties[container_ref.space][container_ref.external_id] == properties
-
-        assert_eventually(replacement_is_queryable)
+        upsert_and_assert_record(cognite_client, mutable_stream, container_ref, sources, ingested_records)
 
 
 class TestStreamsIntegration:
